requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
google-generativeai==0.3.2
textblob==0.17.1
python-dotenv==1.0.0
gunicorn==21.2.0
